+++ modules/linear_boltzmann_solvers/lbs_problem/io/flux_moments_io.cc
@@ -79,7 +79,7 @@ LBSSolverIO::WriteFluxMoments(
   values.reserve(num_local_dofs);
   for (const auto& cell : grid->local_cells)
     for (uint64_t i = 0; i < discretization.GetCellNumNodes(cell); ++i)
-      for (uint64_t m = 0; m < num_moments; ++m)
+      for (unsigned int m = 0; m < num_moments; ++m)
         for (unsigned int g = 0; g < num_groups; ++g)
         {
           const auto dof_map = discretization.MapDOFLocal(cell, i, uk_man, m, g);
@@ -109,7 +109,7 @@ LBSSolverIO::ReadFluxMoments(LBSProblem& lbs_problem,
   log.Log() << "Reading flux moments from " << file_base;
 
   // Read the macro data
-  uint64_t file_num_moments = 0;
+  unsigned int file_num_moments = 0;
   unsigned int file_num_groups = 0;
   uint64_t file_num_local_cells = 0;
   uint64_t file_num_local_nodes = 0;
@@ -205,7 +205,7 @@ LBSSolverIO::ReadFluxMoments(LBSProblem& lbs_problem,
     const uint64_t cell_global_id = file_cell_ids[c];
     const auto& cell = grid->cells[cell_global_id];
     for (uint64_t i = 0; i < discretization.GetCellNumNodes(cell); ++i)
-      for (uint64_t m = 0; m < num_moments; ++m)
+      for (unsigned int m = 0; m < num_moments; ++m)
         for (unsigned int g = 0; g < num_groups; ++g)
         {
           const auto& imap = file_cell_nodal_mapping.at(cell_global_id).at(i);
